+++ artifact/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Artifact, Author
+# Register your models here.
+admin.site.register(Artifact)
+admin.site.register(Author)
