+++ monument/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+from .models import  Author, Monument
+# Register your models here.
+
+admin.site.register(Author)
+admin.site.register(Monument)
