+++ museum/admin.py
@@ -3,9 +3,6 @@ Registration of models to web admin page
 """
 from django.contrib import admin
 
-from .models import Museum, Artifact, Author, Monument
+from .models import Museum
 
 admin.site.register(Museum)
-admin.site.register(Artifact)
-admin.site.register(Author)
-admin.site.register(Monument)
